Changed the performance summary plots to use cmsgraph.cern.ch
Now the graphs are generated from the information stored in
graphite. They don't need jenkins anymore.

The file performance-summary-plots-list contains a list of the
workfows, steps and parameters that we want to show.

The script generate-json-performance-charts reads this list and
generates a json file with the correct structure and urls for
their visualizations.