The release manager script.
The idea is to automate the job of categorizing errors in the
IBs and reporting them as github issues, possibly notifying people
when it's possible to understand the origin of the error.

The script is smart enough to keep track of when issues get fixed
partially (i.e. for a subset of the workflows) and updates the issue
accordingly.