Remove myself from DQM L2s.
After a bit more than two and a half years in Core DQM my last day at CERN has come. Over this time, we changed and improved many things in DQM, and I hope it is now in a better shape than ever and ready to handle the data taking for Run3. I am sure @jfernan2 and @andrius-k will take good care of it.

Thanks Everyone!